fb+frontend: switch framebuffer storage and wire to RGBA
Internal pixel storage in the X11 server framebuffer flips from BGRA
to RGBA, matching what the macOS sidecar already emits, what the
frontend's canvas ImageData layout expects, and what tiny-skia
operates on directly. The frontend is no longer responsible for a
per-pixel BGRA→RGBA swap.

The X11 *wire* format on a little-endian visual is still BGRA, so
PutImage/GetImage and SHM ingest/egress paths swap channels at the
boundary via new `swap_br_in_place` / `swap_br` helpers in the
framebuffer module. A pair of `read_pixel` / `write_pixel` helpers
encapsulate the byte-order-aware pixel access, replacing dozens of
hand-rolled `[B, G, R, A]` shifts throughout image, shm, render,
and drawing handlers.

RENDER's `composite_pixel` and `decode_pixel_bgra` keep their
logical (B, G, R, A) tuple convention but their byte layout
flips to write/read RGBA storage. `resolve_source_pixels` keeps
its BGRA-byte-order *intermediate-buffer* contract so the
composite/glyph paths don't need to change; the cursor handler
in `render::picture` swaps once before sending to the frontend.

Also fixes the pre-existing macOS color-flip bug as a side effect:
the macOS sidecar already produced RGBA output, so removing the
frontend's swap aligns both ingestion paths.

Pre-existing macOS-only build failures (`Color3b`/`Normal3b` taking
`i8` and `SYS_memfd_create` being Linux-only) are fixed in passing
so local cargo check on macOS now works.

+++ crates/sidecar/src/osmesa/gl_bindings.rs
@@ -488,7 +488,7 @@ pub fn gl_rects(x1: i16, y1: i16, x2: i16, y2: i16) {
 
 pub fn gl_color3b(r: i8, g: i8, b: i8) {
     unsafe {
-        gl_generated::Color3b(r as u8, g as u8, b as u8);
+        gl_generated::Color3b(r, g, b);
     }
 }
 pub fn gl_color3d(r: f64, g: f64, b: f64) {
@@ -518,7 +518,7 @@ pub fn gl_color3us(r: u16, g: u16, b: u16) {
 }
 pub fn gl_color4b(r: i8, g: i8, b: i8, a: i8) {
     unsafe {
-        gl_generated::Color4b(r as u8, g as u8, b as u8, a as u8);
+        gl_generated::Color4b(r, g, b, a);
     }
 }
 pub fn gl_color4d(r: f64, g: f64, b: f64, a: f64) {
@@ -594,7 +594,7 @@ pub fn gl_clear_index(c: f32) {
 
 pub fn gl_normal3b(nx: i8, ny: i8, nz: i8) {
     unsafe {
-        gl_generated::Normal3b(nx as u8, ny as u8, nz as u8);
+        gl_generated::Normal3b(nx, ny, nz);
     }
 }
 pub fn gl_normal3d(nx: f64, ny: f64, nz: f64) {
